refactor: exception 한곳에서 처리
