+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -4,7 +4,6 @@ import javax.validation.Valid;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.dto.StationRequest;
 import nextstep.subway.station.dto.StationResponse;
-import nextstep.subway.station.exception.StationNotFoundException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -35,11 +34,6 @@ public class StationController {
     public ResponseEntity<?> deleteStation(@PathVariable Long id) {
         stationService.deleteStationById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler(StationNotFoundException.class)
-    public ResponseEntity<?> handleStationNotFoundException(StationNotFoundException exception) {
-        return ResponseEntity.notFound().build();
     }
 
 }
